requirements.txt
llama-index
openai
chromadb
llama-index-vector-stores-chroma
python-dotenv
pypdf
uvicorn
fastapi
docx2txt